endra setup.sh på lik måte som med humit-språkidentifikator

+++ setup.sh
@@ -1,6 +1,11 @@
 #!/bin/sh
-cat models/sentence_segmentation/x* > models/sentence_segmentation/pytorch_model.bin
-cat models/classification/x* > models/classification/pytorch_model.bin
-cat models/tokenization/x* > models/tokenization/pytorch_model.bin
-rm models/*/x*
+
+ls models/*/x* &>/dev/null
+if [[ "$?" -eq 0 ]]; then
+    cat models/sentence_segmentation/x* > models/sentence_segmentation/pytorch_model.bin
+    cat models/classification/x* > models/classification/pytorch_model.bin
+    cat models/tokenization/x* > models/tokenization/pytorch_model.bin
+    rm models/*/x*
+fi
+
 pip3 install -r requirements.txt
